float scaling

+++ src/options.h
@@ -30,8 +30,8 @@ public:
   QString theme() const;
   void setTheme(QString value);
 
-  int themeScalingFactor() const;
-  void setThemeScalingFactor(int value);
+  float themeScalingFactor() const;
+  void setThemeScalingFactor(float value);
 
   // Returns the blip rate from config.ini (once per X symbols)
   int blipRate() const;
